Move plan purchase features to their own files
