+++ client/blocks/plan-purchase-features/google-analytics-stats.jsx
@@ -1,0 +1,24 @@
+/**
+ * External dependencies
+ */
+import React from 'react';
+import { localize } from 'i18n-calypso';
+
+/**
+ * Internal dependencies
+ */
+import PurchaseDetail from 'components/purchase-detail';
+
+export default localize( ( { selectedSite, translate } ) => {
+	return (
+		<div className="plan-purchase-features__item">
+			<PurchaseDetail
+				icon="stats-alt"
+				title={ translate( 'Stats from Google Analytics' ) }
+				description={ translate( 'Connect to Google Analytics for the perfect complement to WordPress.com stats.' ) }
+				buttonText={ translate( 'Connect Google Analytics' ) }
+				href={ '/settings/analytics/' + selectedSite.slug }
+			/>
+		</div>
+	);
+} );
